+++ crates/cheatnet/src/cheatcodes/spoof.rs
@@ -11,6 +11,12 @@ pub struct TxInfoMock {
     pub transaction_hash: Option<Felt252>,
     pub chain_id: Option<Felt252>,
     pub nonce: Option<Felt252>,
+    pub resource_bounds: Option<Vec<Felt252>>,
+    pub tip: Option<Felt252>,
+    pub paymaster_data: Option<Vec<Felt252>>,
+    pub nonce_data_availabilty_mode: Option<Felt252>,
+    pub fee_data_availabilty_mode: Option<Felt252>,
+    pub account_deployment_data: Option<Vec<Felt252>>,
 }
 
 impl CheatnetState {
@@ -25,6 +31,12 @@ impl CheatnetState {
         transaction_hash: Option<Felt252>,
         chain_id: Option<Felt252>,
         nonce: Option<Felt252>,
+        resource_bounds: Option<Vec<Felt252>>,
+        tip: Option<Felt252>,
+        paymaster_data: Option<Vec<Felt252>>,
+        nonce_data_availabilty_mode: Option<Felt252>,
+        fee_data_availabilty_mode: Option<Felt252>,
+        account_deployment_data: Option<Vec<Felt252>>,
     ) {
         let tx_info = TxInfoMock {
             version,
@@ -34,6 +46,12 @@ impl CheatnetState {
             transaction_hash,
             chain_id,
             nonce,
+            resource_bounds,
+            tip,
+            paymaster_data,
+            nonce_data_availabilty_mode,
+            fee_data_availabilty_mode,
+            account_deployment_data,
         };
         start_cheat(
             &mut self.global_spoof,